snap, snap/ltstrack: keep LTS map mock in export_test.go
