+++ snap/info.go
@@ -2118,19 +2118,6 @@ func SnapdLTSTrackMapFromSnapFile(snapf Container) (trackMap map[int]map[string]
 // SNAPD_LTS_TRACKS; the map lives in the snapd snap. If SNAPD_LTS_TRACKS is
 // absent or empty, trackMap is nil.
 func SnapdLTSTrackMapFromCurrentSnapd() (trackMap map[int]map[string]string, snapdVersion string, err error) {
-	return snapdLTSTrackMapFromCurrentSnapd()
-}
-
-var snapdLTSTrackMapFromCurrentSnapd = snapdLTSTrackMapFromCurrentSnapdImpl
-
-// snapdLTSTrackMapFromCurrentSnapdImpl reads the SNAPD_LTS_TRACKS key from the info
-// file next to this process's snapd binary. Distro packages do not ship this
-// key; it is present when executing from the snapd snap. The directory is
-// snapdtool.InternalLibExecDir(): dirs.DistroLibExecDir for a distro package
-// (/usr/lib/snapd or /usr/libexec/snapd), or the libexec dir of a re-execed
-// snapd snap. This is not the same as /usr/lib/snapd/info inside a snap
-// container (see SnapdLTSTrackMapFromSnapFile).
-func snapdLTSTrackMapFromCurrentSnapdImpl() (trackMap map[int]map[string]string, snapdVersion string, err error) {
 	dir, err := snapdtool.InternalLibExecDir()
 	if err != nil {
 		return nil, "", err
